- enable unmounting of snapshots when we recursively destroy a parent filesystem.
The fix is in the userland and no changes have been made in the kernel. All the
code to traverse the mounted child filesystem and unmount them already exists. It was
ignoring snapshots presumably because this is handled in the kernel in solaris. Since
this fix is much simpler than the kernel change we prefer this.