Protect module automake substitutions with quotes.
"configure --with-config=user" creates a module/Makefile that
contains syntax errors like:

  clean:
    $(MAKE) -C SUBDIRS=`pwd` $@
    if [ -f ]; then $(RM) ; fi

This matters because the top Makefile still calls module/Makefile
even if the kernel modules are not configured for building.

This happens because @LINUX_OBJ@ and @SPL_SYMBOLS@ are undefined if
config/kernel.m4 is not included by the configure script.  This
change results in:

  clean:
    $(MAKE) -C '' SUBDIRS=`pwd` $@
    if [ -f '' ]; then $(RM) ''; fi

which produces better error messages.