Make tgx_sync_thread zio's async
The majority of the recursive operations performed by the dsl
are done either in the context of the tgx_sync_thread or during
pool import.  It is these recursive operations which contribute
greatly to the stack depth.  When this recursion is coupled with
a synchronous I/O in the same context overflow becomes possible.

Previously to handle this case I have focused on keeping the
individual stack frames as light as possible.  This is a good
idea as long as it can be done in a way which doesn't overly
complicate the code.  However, there is a better solution.

If we treat all zio's issued by the tgx_sync_thread as async then
we can use the tgx_sync_thread stack for the recursive parts, and
the zio_* threads for the I/O parts.  This effectively doubles our
available stack space with the only drawback being a small delay
to schedule the zio_* thread.  However, in practice the scheduling
time is so much smaller than the actual I/O time this isn't an
issue.  Another benefit of making the zio async is that the zio
pipeline is now parallel.  That should mean for CPU intensive
pipelines such as compression or dedup performance may be improved.
Initial testing of this change has show a slight decrease in
scrub/resilver time.

With this change in place the worst case stack usage I have so
far observed in 5488 bytes.  This is down several kilobytes from
prior to the change and leaves ample headroom.  This change should
also allow us to revent commit 6656bf5 which did some horrible
things to the traverse_visitbp() callpath in the name of saving
stack.