Add a mountall option to the mount.zfs utility.
Ubuntu systems misbehave if the init stack depends on any mount point
that is not controlled by the /etc/init/mountall.conf upstart job.
Race conditions on the the "local-filesystems" and "mounted" events
are particularly noticeable when when /usr or /var are in ZFS and
mounted by a regular init.d script.

Adding a legacy fstab entry for each ZFS filesystem is onerous and
defeats automatic management features.  Adding a mountall option to
the /sbin/mount.zfs helper allows /sbin/mountall to manage all ZFS
filesystems with just one fstab entry like this:

  # /etc/fstab
  zfs  zfs  zfs  mountall  0  0

This method is an alternative to providing a mountall package that
recognizes ZFS natively, which has unappealing consequences for
maintenance costs and upgrade compatibility.