Remove /etc/zfs/zdev.conf from the install list.
Remove the deprecated `zdev.conf` file from the zfsutils manifest in
anticipation of adding a `vdev_id.conf` converter and/or a debconf test.

Git doesn't track empty directories, so also ensure that `/etc/zfs`
stays registered to the zfsutils package by running `mkdir -p` on it
in the rules file.