Add Tag-zfs-0.6.1.patch
The GA build as published at the PPA and archive.zfsonlinux.org is
missing the final META update that removes the "rc14" string from the
version number. Add the META update to the patch-queue because
revising an orig tarball is disruptive.

This happened because GA was built from a tag using the old 0.6.0.x
workflow instead of from the pristine-tar branch.