Demote the source format and delete the rpm tree.
Commit cb79a4e8bb19908235d3cb646506530457341b76 causes this build error:

	dpkg-source: error: cannot represent change to pkg-zfs/rpm/fedora/zfs-kmod.spec.in:
	dpkg-source: error:   new version is symlink to ../generic/spl-kmod.spec.in
	dpkg-source: error:   old version is plain file
	dpkg-source: error: unrepresentable changes to source

This seems to be an artifact of the security change described at:

  https://launchpad.net/debian/+source/dpkg/1.14.29

Various source package management tools do not update the patch-queue in a way
that satisfies dpkg-source for differential builds, so delete the affected
subtree, which is unused on DEB systems.

This commit is a policy violation that should be reverted after a new upstream
tarball is released.