Re-enable zfs-dracut packaging
This was disabled with f044301bacbe69848fd4c0fcd02f5b4d08d7f298
in January 2012 as it was broken back then, but the situation
has improved vastly in the meantime. The zfs-dracut package
not only works well with dracut-038 on Debian Jessie but offers
one significant advantage over initramfs-tools: The ability to
export a ZFS root on shutdown, eliminating the necessity to
force-import on every reboot.