Move META handling from debian/rules to autoconf.
Let the config/zfs-meta.m4 script ensure version consistency between the META
and debian/changelog files. This change depends on an earlier commit in the
patch-queue branch:  Check-for-META-and-DCH-consistency-in-autoconf.patch

Modifying the META file at build time could cause dpkg-source to fail on an
automatic patch when the recleaning rules glitched. This would randomly
frustrate people trying to recompile the package in a non-pristine build area.