Subvert Launchpad orig versioning.
The buildpackage tools are incapable of doing differential snapshot
builds at Launchpad because the distance between the zfs-0.6.3
release tag and the current HEAD is large.

Create a bogus upstream zfs-0.6.3.0 tag so that the orig tarball is
regenerated and snapshots can be built for the V series, which is now
in beta.